Rearrange dual-axis input processors
